Made Particle struct a clas and moved it to separate .hpp and .cpp files. Added files for QuadTree class for new particle collision detection system. TODO: Implement QuadTree Class
