+++ Quadtree.hpp
@@ -1,0 +1,8 @@
+#ifndef QUADTREE
+#define QUADTREE
+
+class QuadTree {
+
+};
+
+#endif
